requirements.txt
asgiref==3.3.1
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
importlib-metadata==3.4.0
Markdown==3.3.3
PyJWT==1.7.1
pytz==2021.1
simplejson==3.17.2
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
